% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_sex}
\alias{age_sex}
\title{Age and sex demographics by state}
\format{
Age and sex demographics by state
\describe{
\item{State}{US States}
\item{Sex}{Gender}
\item{LBound}{Lower bound of age range}
\item{UBound}{Upper bound of age range}
\item{Population}{Population with applicable attributes}
}
}
\usage{
age_sex
}
\description{
Age and sex demographics by state
}
\keyword{datasets}
