% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_fields}
\alias{data_fields}
\title{Default and available fields and states}
\format{
A list of field and state names
\describe{
\item{available_fields}{available fields in the CDC data}
\item{available_states}{available names of the US territory in the CDC data}
\item{default_fields}{default fields specified for all functions in the package}
\item{default_states}{default states specified for all functions in the package}
}
}
\usage{
data_fields
}
\description{
Default and available fields and states
}
\keyword{datasets}
