% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_importer.R
\name{format_covid_data}
\alias{format_covid_data}
\title{formatting covid data}
\usage{
format_covid_data(df)
}
\arguments{
\item{df}{}
}
\value{
a data.frame
}
\description{
formatting covid data
}
\keyword{internal}
