% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_importer.R
\name{generate_url}
\alias{generate_url}
\title{Generate URL}
\usage{
generate_url(
  start_date,
  end_date,
  states,
  fields,
  consent_cases,
  consent_deaths,
  limit
)
}
\arguments{
\item{start_date}{start date}

\item{end_date}{end date}

\item{states}{states}

\item{fields}{fields}

\item{consent_cases}{consent cases}

\item{consent_deaths}{consent deaths}

\item{limit}{limit}
}
\value{
a string
}
\description{
Generate URL
}
