% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{Insurance characteristics by population}
\format{
Insurance characteristics by state
\describe{
\item{State}{available fields in the CDC data}
\item{year}{Self explanatory}
\item{agecat}{Self explanatory}
\item{racecat}{Self explanatory}
\item{sexcat}{Self explanatory}
\item{iprcat}{Self explanatory}
\item{Number in Demographic Group}{Self explanatory}
\item{Number Uninsured}{Self explanatory}
\item{Number Insured}{Self explanatory}
\item{Percent uninsured in demographic group for income category}{Self explanatory}
\item{Percent insured in demographic group for income category}{Self explanatory}
\item{Percent uninsured in demographic group for all income levels}{Self explanatory}
\item{Percent insured in demographic group for all income levels}{Self explanatory}
}
}
\usage{
insurance
}
\description{
Insurance characteristics by population
}
\keyword{datasets}
