% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_importer.R
\name{pull_covid_data}
\alias{pull_covid_data}
\title{Pull COVID data from CDC}
\usage{
pull_covid_data(
  start_date,
  end_date,
  states = NULL,
  fields = NULL,
  consent_cases = TRUE,
  consent_deaths = TRUE,
  limit = 1e+05
)
}
\arguments{
\item{start_date}{start date}

\item{end_date}{end date}

\item{states}{this input argument can be (1) \code{"default"}, which uses a vector of default vector of states from the
internal data \code{data_fields}; (2) a vector of state abbreviations specified
by the user, e.g., \code{c("GA", "FL")}.}

\item{fields}{a vector of fields, e.g., \code{c("submission_date", "state")}. If nothing is provided
the function extracts the default vector of fields from the internal data \code{data_fields}.}

\item{consent_cases}{consent cases}

\item{consent_deaths}{consent deaths}

\item{limit}{limit of records}
}
\description{
Pull COVID data from CDC
}
\examples{
df_filtered_1 <- pull_covid_data("2020-09-30", "2020-10-30")
df_filtered_2 <- pull_covid_data("2020-09-30", "2020-10-30", limit=100)
df_filtered_3 <- pull_covid_data("2020-01-01", "2020-10-30",
                 states=c("FL", "TX", "CA"),
                 fields=c("state", "new_case", "new_death", "tot_cases", "tot_death", "submission_date"),
                 limit=100)

}
