% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_ethnicity}
\alias{race_ethnicity}
\title{Race demographics by state}
\format{
Race demographics by state
\describe{
\item{State}{US State}
\item{Race}{Race of population}
\item{Population}{Population count}
}
}
\usage{
race_ethnicity
}
\description{
Race demographics by state
}
\keyword{datasets}
