% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_graphs.R
\name{covid_deaths_graph}
\alias{covid_deaths_graph}
\title{Plot deaths}
\usage{
covid_deaths_graph(start_date, end_date, per100k = TRUE)
}
\arguments{
\item{start_date}{start date}

\item{end_date}{end date}

\item{per100k}{per100k}
}
\value{
a ggplot object
}
\description{
Plot deaths
}
\examples{
covid_deaths_graph("2020-03-30", "2020-08-30",  per100k = TRUE)
covid_deaths_graph("2020-03-30", "2020-08-30",  per100k = FALSE)

}
