% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_ts_forecast.R
\name{covid_ts_forecast}
\alias{covid_ts_forecast}
\title{Forecast COVID-19 new cases or new deaths in USA from now up to a maximum of 100 days}
\usage{
covid_ts_forecast(
  start_date = "2020-01-22",
  fvar = c("new_case", "new_death"),
  state.filter = NULL,
  region.filter = NULL,
  model = c("ets", "bats", "tbats", "auto.arima"),
  pred.days = 30
)
}
\arguments{
\item{start_date}{Start date. Default uses 2020-01-22.}

\item{fvar}{Forecasting object (new_case or new_death).}

\item{state.filter}{Optional vector specifying a subset of states to be used. This cannot be used in conjunction with region.filter. Default uses c("all") i.e., forecast will be done for USA.}

\item{region.filter}{Optional vector specifying a subset of regions to be used from a list c("Northeast","Midwest","South","West"). This cannot be used in conjunction with state.filter. Default uses state.filter=c("all") i.e., forecast will be done for USA.}

\item{model}{Forecasting model to be used from a list of c('naive','ets','bats','tbats','auto.arima'). Default uses "auto.arima"}

\item{pred.days}{Number of days for forecasting. Limited to 100 days. Default uses 30 days.}
}
\description{
Forecast COVID-19 new cases or new deaths in USA from now up to a maximum of 100 days
}
\details{
\describe{
\item{start_date}{Input data is obtained from pull_covid_data function. The source data is from CDC website. Data is available from 2020-01-22. }
\item{fvar}{Only one object can be used at a time.}
\item{state.filter}{2 character codes for states should be used. If not specified, state.filter will calculate sum of new cases (new_case) or new deaths (new_death) across each day}
\item{region.filter}{If this is left blank along with state.filter then state.filter=c("all") will be used}
\item{model}{Only a single model can be used at a time.}
}
}
\examples{
covid_ts_forecast(start_date='2020-01-01', fvar="new_case", state.filter=c("MA","CA"), pred.days = 30)
covid_ts_forecast(fvar="new_death", state.filter=c("all"), pred.days = 60)
covid_ts_forecast(start_date='2020-02-01', fvar="new_death", model=c("auto.arima"), pred.days = 60)
covid_ts_forecast(start_date='2020-04-30', fvar="new_case", region.filter=c("Midwest","South"), model=c("tbats"), pred.days=40)
covid_ts_forecast(fvar="new_death")

}
